+++ castle/cms/browser/viewlets/socialtags.py
@@ -1,3 +1,4 @@
+from plone import api
 from plone.memoize.view import memoize
 from castle.cms.browser.syndication import FolderFeed
 from plone.app.layout.viewlets.social import SocialTagsViewlet as BaseSocialTagsViewlet
@@ -13,21 +14,27 @@ class SocialTagsViewlet(BaseSocialTagsViewlet):
         tags = super(SocialTagsViewlet, self)._get_tags()
 
         site = getSite()
+        site_title = api.portal.get_registry_record('plone.site_title', default=None)
         feed = FolderFeed(site)
         item = queryMultiAdapter((self.context, feed), IFeedItem, default=None)
         if item is None:
             item = DexterityItem(self.context, feed)
 
-        if item.has_image:
-            for tag in tags[:]:
+        for tag in tags:
+            if site_title and (tag.get('property', '') == 'og:title' or
+                                    tag.get('name', '') == 'twitter:title'):
+                tag['content'] = '{} | {}'.format(tag['content'], site_title)
+
+            if item.has_image:
                 if ('image' in tag.get('property', '') or
                         'image' in tag.get('itemprop', '') or
                         'image' in tag.get('name', '')):
                     tags.remove(tag)
-                elif tag.get('name') == 'twitter:card':
+                elif tag.get('name', '') == 'twitter:card':
                     # change to large summary
                     tag['content'] = 'summary_large_image'
 
+        if item.has_image:
             tags.extend([
                 dict(property="og:image", content=item.image_url),
                 dict(itemprop="image", content=item.image_url),